fis and add numbe rof units
